refactor: clean project structure and dependencies
