+++ cmd/docker-gen/main.go
@@ -10,7 +10,7 @@ import (
 
 	"github.com/BurntSushi/toml"
 	docker "github.com/fsouza/go-dockerclient"
-	"github.com/jwilder/docker-gen"
+	"github.com/nginx-proxy/docker-gen/internal/dockergen"
 )
 
 type stringslice []string
