+++ go.mod
@@ -1,27 +1,9 @@
-module github.com/jwilder/docker-gen
+module github.com/nginx-proxy/docker-gen
 
-go 1.11
+go 1.16
 
 require (
-	github.com/Azure/go-ansiterm v0.0.0-20170929234023-d6e3b3328b78 // indirect
 	github.com/BurntSushi/toml v0.3.1
-	github.com/Microsoft/go-winio v0.4.16 // indirect
-	github.com/Nvveen/Gotty v0.0.0-20120604004816-cd527374f1e5 // indirect
-	github.com/containerd/continuity v0.0.0-20210315143101-93e15499afd5 // indirect
-	github.com/docker/docker v1.4.2-0.20171014114940-f2afa2623594 // indirect
-	github.com/docker/go-connections v0.4.0 // indirect
-	github.com/docker/go-units v0.3.2 // indirect
-	github.com/fsouza/go-dockerclient v0.0.0-20171009031830-d2a6d0596004
-	github.com/gogo/protobuf v1.3.2 // indirect
-	github.com/google/go-cmp v0.5.5 // indirect
-	github.com/gorilla/context v1.1.1 // indirect
-	github.com/gorilla/mux v0.0.0-20160718151158-d391bea3118c // indirect
-	github.com/gotestyourself/gotestyourself v2.2.0+incompatible // indirect
-	github.com/opencontainers/image-spec v1.0.1 // indirect
-	github.com/opencontainers/runc v0.1.1 // indirect
-	github.com/opencontainers/selinux v1.8.0 // indirect
-	github.com/sirupsen/logrus v1.8.1 // indirect
+	github.com/fsouza/go-dockerclient v1.7.2
 	github.com/stretchr/testify v1.7.0
-	golang.org/x/sys v0.0.0-20210331175145-43e1dd70ce54 // indirect
-	gotest.tools v2.2.0+incompatible // indirect
 )
