+++ internal/dockergen/context.go
